requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy[asyncio]==2.0.35
asyncpg==0.29.0
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
itsdangerous==2.2.0

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
slowapi==0.1.9  # Rate limiting

# Валидация и санитизация
bleach==6.1.0

# Мониторинг и логирование
python-json-logger==2.0.7
psutil==6.1.0  # System metrics для health check
sentry-sdk[fastapi]==2.19.2  # Error tracking

# Push-уведомления
firebase-admin==7.1.0  # FCM push notifications

# AI и обработка файлов
openai==1.57.4  # OpenAI API для генерации вопросов
pypdf==5.1.0  # Извлечение текста из PDF
python-docx==1.1.2  # Извлечение текста из DOCX

# QR-генератор
qrcode[pil]==8.0
Pillow==10.4.0
